+++ src/main/java/terracraft/mixin/world/BiomeManagerMixin.java
@@ -17,7 +17,7 @@ import terracraft.common.utility.CorruptionHelper;
 @Mixin(BiomeManager.class)
 public class BiomeManagerMixin {
 
-    //todo: currently can cause a crash randomly when the player dies, probably need to have set to corruption biome if surrounded by other corruption blocks
+    //todo: currently can cause a crash randomly, has to do with block tint for grass and foliage, ClientLevelMixin may temporarily fix this
     @Inject(at = @At("RETURN"), method = "getBiome", cancellable = true)
     public void spreadCorruptionBiome(BlockPos pos, CallbackInfoReturnable<Holder<Biome>> info) {
         Minecraft mc = Minecraft.getInstance();