+++ CHANGELOG.md
@@ -7,13 +7,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [1.3.0] - 2022-04-29
 ### Added
 - Demonite Ore Generation (Corruption Biomes Only)
-- Corruption Biome Pits
+- Corruption Biome Pits (WIP)
+- Movement Ordering System (So that Cloud in a Bottle and Rocket Boots don't happen at the same time)
 
 ### Changed
 - Various improvements to Demon Eyes
 - Removed jump height increase while sprinting and double jumping
 - Accessory effects are now disabled in creative mode
 - Information Accessories effects can now be toggled in inventory
+
+### Fixed
+- Random Corruption spread crash (maybe, at the cost of block tint gradient in corruption biome)
 
 ## [1.2.0] - 2022-04-26
 ### Added